use drain and create batches

+++ bottlecap/src/metrics/constants.rs
@@ -5,12 +5,12 @@ pub const CONTEXTS: usize = 1024;
 
 pub static MAX_CONTEXTS: usize = 65_536; // 2**16, arbitrary
 
-const MB: usize = 1_024 * 1_024;
+const MB: u64 = 1_024 * 1_024;
 
 pub(crate) const MAX_ENTRIES_SINGLE_METRIC: usize = 1_000;
 
-pub(crate) const MAX_SIZE_BYTES_SINGLE_METRIC: usize = 5 * MB;
+pub(crate) const MAX_SIZE_BYTES_SINGLE_METRIC: u64 = 5 * MB;
 
 pub(crate) const MAX_ENTRIES_SKETCH_METRIC: usize = 1_000;
 
-pub(crate) const MAX_SIZE_SKETCH_METRIC: usize = 62 * MB;
+pub(crate) const MAX_SIZE_SKETCH_METRIC: u64 = 62 * MB;
